refactor: split memo funciton to component
